Converted custom commands and hooks to TypeScript.
Typescript can't handle dynamic imports on its own. Using SystemJS to handle this.

The game_id variable (currently set in index.html but eventually to be set by the back end) controls which adventure gets imported by System.import. No matter which one it is, it's called "adventure". This makes it possible to "import" files from that adventure into the core Typescript files.