Fix exact matching of artifact/monster names
The matching logic gives preference to exact artifact/monster name matches (e.g., if there is a "silver chest" and a "chest", saying "open chest" will open the regular "chest" and you have to say "open silver chest" to open the other one). This prevents an artifact with a longer name from "hiding" the one with a shorter name.